Add an intrinsic to communicate the parameter struct to the compiler.
In order to do the necessary optimizations in the compiler to avoid
accessing members of the parameter struct with loads throughout the
program, the compiler must know that the functor object is immutable and
thread invariant for the duration of the parallel region.

This commit adds an intrinsic that can communicate this to the
NextSilicon compiler.